Refactor and enhance agent functionality with new group chat and MCP integration
- Removed obsolete remote code execution script (08_generate_run_code_in_remote_container_on_aca_langchain.py).
- Added group chat functionality for collaborative problem-solving (09_group_chat_coding_problem.py).
- Implemented restaurant booking agent with MCP integration for enhanced user interaction (10_agent_to_mcp.py).
- Created GitHub issue querying agent using MCP server (11_mcp_gh_client.py).
- Updated settings to include kernel setup for chat completion service.
- Improved error handling and logging throughout the new implementations.
