+++ src/10_agent_to_mcp.py
@@ -44,42 +44,20 @@ from pathlib import Path
 
 import dotenv
 from semantic_kernel.agents import ChatCompletionAgent, ChatHistoryAgentThread
-from semantic_kernel.connectors.ai.open_ai import AzureChatCompletion
 from semantic_kernel.connectors.mcp import MCPStdioPlugin
 from semantic_kernel.core_plugins.time_plugin import TimePlugin
-from semantic_kernel.kernel import Kernel
 
-from settings import llm_config
+import settings
 
 dotenv.load_dotenv()
 
-azure_openai_endpoint = os.getenv("AZURE_OPENAI_URL")
-
-
-def setup_chat_service(kernel: Kernel, service_id: str) -> None:
-    """Set up a chat completion service for the kernel."""
-    deployment_name = llm_config.get("config", {}).get("model", "gpt-4o")
-    endpoint = llm_config.get("config", {}).get(
-        "azure_endpoint", azure_openai_endpoint
-    )
-    api_key = llm_config.get("config", {}).get("api_key", None)
-    api_version = llm_config.get("config", {}).get("api_version", "2024-06-01")
-
-    chat_service = AzureChatCompletion(
-        service_id=service_id,
-        endpoint=endpoint,
-        api_key=api_key,
-        api_version=api_version,
-        deployment_name=deployment_name,
-    )
-    kernel.add_service(chat_service)
-
 
 async def main():
-    # 1. Create the agent
-    kernel = Kernel()
+    # 1. Create the kernel for the agent
     service_id = "restaurant-agent"
-    setup_chat_service(kernel, service_id)
+    kernel = settings.setup_kernel_with_chat_completion_service(
+        service_id=service_id
+    )
 
     async with (
         MCPStdioPlugin(
